security-tool/requirements.txt
zaproxy
requests
python-dotenv